+++ src/main/java/baseball/model/ball/BallMaker.java
@@ -1,4 +1,4 @@
-package baseball.model;
+package baseball.model.ball;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.ArrayList;
