+++ src/test/java/baseball/model/BallMakerTest.java
@@ -2,6 +2,8 @@ package baseball.model;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import baseball.model.ball.Ball;
+import baseball.model.ball.BallMaker;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
