+++ src/test/java/baseball/model/judge/BaseballJudgeTest.java
@@ -2,7 +2,7 @@ package baseball.model.judge;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import baseball.model.Ball;
+import baseball.model.ball.Ball;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
